+++ src/volumito/__init__.py
@@ -24,6 +24,9 @@ from volumito.clients import (
     Queue,
     QueueTrack,
     RemoteCommandResult,
+    SearchResultItem,
+    SearchResultList,
+    SearchResults,
     Story,
     SuccessResponse,
     SystemInfo,
@@ -48,7 +51,7 @@ from volumito.clients import (
     remote_music_path,
 )
 
-__version__ = "0.0.37"
+__version__ = "0.0.38"
 __author__ = "Alberto Pettarin"
 __email__ = "alberto@albertopettarin.it"
 
@@ -68,6 +71,9 @@ __all__ = [
     "PushNotification",
     "Queue",
     "QueueTrack",
+    "SearchResultItem",
+    "SearchResultList",
+    "SearchResults",
     "Story",
     "SuccessResponse",
     "SystemInfo",